Extract `mergeSeries` to its own file
